abduco.nix: if /early/missing is missing, have it launch /bin/sh
If any of the /early/{run,finish,fail} scripts are missing, six-initrd will
invoke /early/missing instead.  However, previously, if /early/missing was
itself missing, the system would simply kernel panic due to PID1 dying.

This commit causes six-initrd to create /early/missing if it does not exist,
populating it with a script which invokes an interactive shell.  On systems
where this is undesirable (kiosks, etc) you should ensure that /early/missing
exists and point it at something (like /bin/false) that isn't a shell.